==> notebooks/2_data_preprocessing.ipynb <==
# Data Preprocessing

import numpy as np
import matplotlib.pyplot as plt
import os
import numpy as np
import seaborn as sns
import json
import librosa
import soundfile as sf
import random
import pandas as pd
from pydub import AudioSegment
from tqdm import tqdm
from birdlib import utils

DATASET_NAME = 'dataset'
AUDIO_SOURCE = '/home/giacomoschiavo/Tovanella'

DATASET_PATH = f'/home/giacomoschiavo/segments/{DATASET_NAME}'
TRAIN_PATH = f'/home/giacomoschiavo/segments/{DATASET_NAME}/train'
VALID_PATH = f'/home/giacomoschiavo/segments/{DATASET_NAME}/valid'
TEST_PATH = f'/home/giacomoschiavo/segments/{DATASET_NAME}/test'

species_list = set(os.listdir(TEST_PATH)).intersection(set(os.listdir(TRAIN_PATH)))
if 'Engine_Engine' in species_list:
    species_list.remove('Engine_Engine')
if 'Cuculus canorus_Common Cuckoo' in species_list:
    species_list.remove('Cuculus canorus_Common Cuckoo')

def print_dataset_count_table(dataset_path):
    train_folder = f"{dataset_path}/train"
    valid_folder = f"{dataset_path}/valid"
    test_folder = f"{dataset_path}/test"

    dataset_count = {}
    for species in os.listdir(test_folder):
        if species not in species_list:
            continue
        dataset_count[species] = {
            "train": len(os.listdir(os.path.join(train_folder, species))) if os.path.exists(os.path.join(train_folder, species)) else 0,
            "valid": len(os.listdir(os.path.join(valid_folder, species))) if os.path.exists(os.path.join(valid_folder, species)) else 0,
            "test": len(os.listdir(os.path.join(test_folder, species))) if os.path.exists(os.path.join(test_folder, species)) else 0
        }

    dataset_species_count_df = pd.DataFrame.from_dict(dataset_count, orient='index')
    dataset_species_count_df.index.name = 'Species'
    return dataset_species_count_df.sort_values(by=["train"], ascending=False)


print_dataset_count_table(DATASET_PATH)

print_dataset_count_table(DATASET_PATH)

dates_count = utils.get_date_count(TRAIN_PATH, species_list)

# Test Integration

train_integration, test_integration = utils.split_dataset(dates_count, TEST_PATH, test_ratio=0.2)

with open(f"utils/{DATASET_NAME}/train_integration.json", 'w') as f:
    json.dump(train_integration, f)

with open(f"utils/{DATASET_NAME}/test_integration.json", 'w') as f:
    json.dump(test_integration, f)

# with open(f"utils/{DATASET_NAME}/train_integration.json") as f:
#     train_integration = json.load(f)

# with open(f"utils/{DATASET_NAME}/test_integration.json") as f:
#     test_integration = json.load(f)

utils.move_by_date(test_integration, TRAIN_PATH, TEST_PATH)

# Validation creation

# take 5% of the last samples from each date, samples are ordered by the segment number (32_5 -> 32.5)
for species in species_list:
    os.makedirs(os.path.join(VALID_PATH, species), exist_ok=True)
    train_species_path = os.path.join(TRAIN_PATH, species)
    train_files = sorted(os.listdir(train_species_path))
    date_count = utils.get_date_count(TRAIN_PATH, species_list)
    for date in dates_count[species]:
        date_files = [f for f in train_files if f.startswith(date)]
        date_files = sorted(date_files, key=lambda x: float(x.split('_')[-2] + '.' + x.split('_')[-1].split('.')[0]))
        if len(date_files) == 0:
            continue
        num_samples_to_move = max(1, int(len(date_files) * 0.1))
        samples_to_move = date_files[-num_samples_to_move:]
        
        for sample in samples_to_move:
            src_path = os.path.join(train_species_path, sample)
            dest_path = os.path.join(VALID_PATH, species, sample)
            os.makedirs(os.path.dirname(dest_path), exist_ok=True)
            os.rename(src_path, dest_path)        


print_dataset_count_table(DATASET_PATH)

len(os.listdir(TRAIN_PATH)), len(os.listdir(TEST_PATH)), len(os.listdir(VALID_PATH))

## Removed unused species

valid_species = os.listdir(VALID_PATH)
REMOVED_PATH = f'{DATASET_PATH}/removed'
os.makedirs(REMOVED_PATH, exist_ok=True)
REMOVED_TRAIN_PATH = f'{REMOVED_PATH}/train'
os.makedirs(REMOVED_TRAIN_PATH, exist_ok=True)
for species in os.listdir(TRAIN_PATH):
    if species not in valid_species:
        os.makedirs(os.path.join(REMOVED_TRAIN_PATH, species), exist_ok=True)
        os.rename(
            os.path.join(TRAIN_PATH, species),
            os.path.join(REMOVED_TRAIN_PATH, species)
        )

valid_species = os.listdir(VALID_PATH)
REMOVED_TEST_PATH = f'{REMOVED_PATH}/test'
os.makedirs(REMOVED_TEST_PATH, exist_ok=True)
for species in os.listdir(TEST_PATH):
    if species not in valid_species:
        os.makedirs(os.path.join(REMOVED_TEST_PATH, species), exist_ok=True)
        os.rename(
            os.path.join(TEST_PATH, species),
            os.path.join(REMOVED_TEST_PATH, species)
        )

len(os.listdir(TRAIN_PATH)), len(os.listdir(TEST_PATH)), len(os.listdir(VALID_PATH))

print_dataset_count_table(DATASET_PATH)

## Undo Valid

# # MOVE VALID TO TRAIN
# for species in tqdm(os.listdir(VALID_PATH)):
#     if species not in species_list:
#         continue
#     species_train_path = os.path.join(TRAIN_PATH, species)
#     species_valid_path = os.path.join(VALID_PATH, species)
#     if not os.path.exists(species_train_path):
#         os.makedirs(species_train_path)

#     for file in os.listdir(species_valid_path):
#         src = os.path.join(species_valid_path, file)
#         dst = os.path.join(species_train_path, file)
#         if not os.path.exists(dst):
#             os.rename(src, dst)
#         else:
#             print(f"File {dst} already exists, skipping {src}")
