==> notebooks/7_best_model_selection.ipynb <==
import os
import csv
import json

RESULTS_PATH = "/home/giacomoschiavo/finetuning-BirdNET/models/CustomCNN/"
# load all csv file sin the results path
def load_json_files(path):
    csv_files = [f for f in os.listdir(path) if f.endswith('.json') and 'best_model' not in f and '_0' in f]
    data = {}
    print(csv_files)
    for file_name in csv_files:
        with open(os.path.join(path, file_name), 'r') as f:
            data[file_name] = json.load(f)
    return data

all_results = load_json_files(RESULTS_PATH)

best_results = {file_name: config[0] for file_name, config in all_results.items()}
best_results

import pandas as pd
def analyze_model_results(best_results):
    # Lista per contenere i dati appiattiti per il DataFrame
    processed_data = []

    # Processa ogni elemento in all_results_data
    for file_name, model in best_results.items():
        row = {
            'file_name': file_name,
            'model_id': model.get('model_id'),
            'micro_f1': model.get('micro_f1'),
            'weighted_f1': model.get('weighted_f1'),
            'samples_f1': model.get('samples_f1'),
            'mean_f1': model.get('mean_f1')
        }

        config = model.get('config', {})
        row['num_conv_layers'] = config.get('num_conv_layers')
        row['kernel_sizes'] = str(config.get('kernel_sizes')) # Converti in stringa per display
        row['channels'] = str(config.get('channels'))       # Converti in stringa per display
        row['batch_size'] = config.get('batch_size')
        row['dropout'] = config.get('dropout')
        row['dense_hidden'] = config.get('dense_hidden')
        
        processed_data.append(row)

    # Crea il DataFrame di Pandas
    df = pd.DataFrame(processed_data)

    # Riorganizza le colonne per una migliore leggibilità
    df = df[['file_name', 'model_id', 'num_conv_layers', 'kernel_sizes', 'channels', 
             'batch_size', 'dropout', 'dense_hidden', 
             'micro_f1', 'weighted_f1', 'samples_f1', 'mean_f1']]

    return df, df['mean_f1'].idxmax() # Restituisce il DataFrame e il miglior modello

# Chiama la funzione per analizzare i risultati
df_results, index = analyze_model_results(best_results)


df_results.sort_values(by='mean_f1', ascending=False)

best_model_info = df_results.sort_values(by='mean_f1', ascending=False).iloc[0]

best_model_info.to_json('models/CustomCNN/best_model_config.json')

best_model_info
